+++ docs/client/components/pages/Image/index.js
@@ -41,6 +41,18 @@ export default class App extends Component {
         <Separator />
         <Container>
           <Heading level={2}>Image</Heading>
+          <Heading level={3}>Supported Environment</Heading>
+          <ul className={styles.list}>
+            <li className={styles.listEntry}>
+              Desktop: Yes
+            </li>
+            <li className={styles.listEntry}>
+              Mobile: Yes
+            </li>
+            <li className={styles.listEntry}>
+              Screen-reader: No
+            </li>
+          </ul>
         </Container>
         <AlternateContainer>
           <Heading level={2}>Getting Started</Heading>
@@ -128,7 +140,7 @@ export default class App extends Component {
           <Code code={simpleExampleEditorStylesCode} name="editorStyles.css" />
         </Container>
         <Container>
-          <Heading level={2}>Alignment + Resize + Focus + Drag'n'Drop + Drag'n'Drop Upload Example</Heading>
+          <Heading level={2}>Alignment + Resize + Focus + Drag&apos;n&apos;Drop + Drag&apos;n&apos;Drop Upload Example</Heading>
           <CustomImageEditor />
           <Code code={customExampleCode} name="AddImageEditor.js" />
           <Code code={customExampleEditorStylesCode} name="editorStyles.css" />