+++ draft-js-mention-plugin/src/MentionSuggestionsPortal/index.js
@@ -2,6 +2,13 @@ import React, { Component } from 'react';
 
 export default class MentionSuggestionsPortal extends Component {
 
+  constructor(props) {
+    super(props);
+    // Note: this is a workaround for an obscure issue: https://github.com/draft-js-plugins/draft-js-plugins/pull/667/files
+    // Ideally we can remove this in the future.
+    this.searchPortalRef = (element) => { this.searchPortal = element; };
+  }
+
   // When inputting Japanese characters (or any complex alphabet which requires
   // hitting enter to commit the characters), that action was causing a race
   // condition when we used componentWillMount. By using componentDidMount
@@ -11,6 +18,7 @@ export default class MentionSuggestionsPortal extends Component {
   // was still in the DOM, so it wasn't re-registering the offsetkey.
   componentDidMount() {
     this.props.store.register(this.props.offsetKey);
+    this.props.store.setIsOpened(true);
     this.updatePortalClientRect(this.props);
 
     // trigger a re-render so the MentionSuggestions becomes active
@@ -23,6 +31,7 @@ export default class MentionSuggestionsPortal extends Component {
 
   componentWillUnmount() {
     this.props.store.unregister(this.props.offsetKey);
+    this.props.store.setIsOpened(false);
   }
 
   updatePortalClientRect(props) {
@@ -38,7 +47,7 @@ export default class MentionSuggestionsPortal extends Component {
     return (
       <span
         className={this.key}
-        ref={(element) => { this.searchPortal = element; }}
+        ref={this.searchPortalRef}
       >
         {this.props.children}
       </span>