+++ draft-js-mention-plugin/src/index.js
@@ -12,6 +12,8 @@ import mentionSuggestionsEntryStyles from './mentionSuggestionsEntryStyles.css';
 import suggestionsFilter from './utils/defaultSuggestionsFilter';
 import defaultPositionSuggestions from './utils/positionSuggestions';
 
+export { default as MentionSuggestions } from './MentionSuggestions';
+
 export default (config = {}) => {
   const defaultTheme = {
     // CSS class for mention text
@@ -46,6 +48,7 @@ export default (config = {}) => {
   let searches = Map();
   let escapedSearch;
   let clientRectFunctions = Map();
+  let isOpened;
 
   const store = {
     getEditorState: undefined,
@@ -73,6 +76,9 @@ export default (config = {}) => {
       searches = searches.delete(offsetKey);
       clientRectFunctions = clientRectFunctions.delete(offsetKey);
     },
+
+    getIsOpened: () => isOpened,
+    setIsOpened: (nextIsOpened) => { isOpened = nextIsOpened; },
   };
 
   // Styles are overwritten instead of merged as merging causes a lot of confusion.
@@ -86,6 +92,7 @@ export default (config = {}) => {
     theme = defaultTheme,
     positionSuggestions = defaultPositionSuggestions,
     mentionComponent,
+    mentionSuggestionsComponent = MentionSuggestions,
     entityMutability = 'SEGMENTED',
     mentionTrigger = '@',
     mentionRegExp = defaultRegExp,
@@ -103,7 +110,7 @@ export default (config = {}) => {
     mentionPrefix,
   };
   return {
-    MentionSuggestions: decorateComponentWithProps(MentionSuggestions, mentionSearchProps),
+    MentionSuggestions: decorateComponentWithProps(mentionSuggestionsComponent, mentionSearchProps),
     decorators: [
       {
         strategy: mentionStrategy(mentionTrigger),