+++ draft-js-plugins-editor/src/Editor/index.js
@@ -1,5 +1,6 @@
-/* eslint-disable no-continue */
+/* eslint-disable no-continue,no-restricted-syntax */
 import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import {
   Editor,
   EditorState,
@@ -18,13 +19,13 @@ import * as defaultKeyBindingPlugin from './defaultKeyBindingPlugin';
 class PluginEditor extends Component {
 
   static propTypes = {
-    editorState: React.PropTypes.object.isRequired,
-    onChange: React.PropTypes.func.isRequired,
-    plugins: React.PropTypes.array,
-    defaultKeyBindings: React.PropTypes.bool,
-    defaultBlockRenderMap: React.PropTypes.bool,
-    customStyleMap: React.PropTypes.object,
-    decorators: React.PropTypes.array,
+    editorState: PropTypes.object.isRequired,
+    onChange: PropTypes.func.isRequired,
+    plugins: PropTypes.array,
+    defaultKeyBindings: PropTypes.bool,
+    defaultBlockRenderMap: PropTypes.bool,
+    customStyleMap: PropTypes.object,
+    decorators: PropTypes.array,
   };
 
   static defaultProps = {
@@ -39,17 +40,17 @@ class PluginEditor extends Component {
     super(props);
 
     const plugins = [this.props, ...this.resolvePlugins()];
-    for (const plugin of plugins) {
-      if (typeof plugin.initialize !== 'function') continue;
+    plugins.forEach((plugin) => {
+      if (typeof plugin.initialize !== 'function') return;
       plugin.initialize(this.getPluginMethods());
-    }
+    });
 
     // attach proxy methods like `focus` or `blur`
-    for (const method of proxies) {
+    proxies.forEach((method) => {
       this[method] = (...args) => (
         this.editor[method](...args)
       );
-    }
+    });
 
     this.state = {}; // TODO for Nik: ask ben why this is relevent
   }
@@ -65,11 +66,7 @@ class PluginEditor extends Component {
       .filter((decorator) => this.decoratorIsCustom(decorator));
 
     const multiDecorator = new MultiDecorator(
-      [
-        ...customDecorators,
-        compositeDecorator,
-      ]
-    );
+      customDecorators.push(compositeDecorator));
 
     const editorState = EditorState.set(this.props.editorState, { decorator: multiDecorator });
     this.onChange(moveSelectionToEnd(editorState));
@@ -126,25 +123,21 @@ class PluginEditor extends Component {
   createEventHooks = (methodName, plugins) => (...args) => {
     const newArgs = [].slice.apply(args);
     newArgs.push(this.getPluginMethods());
-    for (const plugin of plugins) {
-      if (typeof plugin[methodName] !== 'function') continue;
-      const result = plugin[methodName](...newArgs);
-      if (result === true) return true;
-    }
 
-    return false;
+    return plugins.some((plugin) =>
+      typeof plugin[methodName] === 'function'
+        && plugin[methodName](...newArgs) === true
+    );
   };
 
   createHandleHooks = (methodName, plugins) => (...args) => {
     const newArgs = [].slice.apply(args);
     newArgs.push(this.getPluginMethods());
-    for (const plugin of plugins) {
-      if (typeof plugin[methodName] !== 'function') continue;
-      const result = plugin[methodName](...newArgs);
-      if (result === 'handled') return 'handled';
-    }
 
-    return 'not-handled';
+    return plugins.some((plugin) =>
+      typeof plugin[methodName] === 'function'
+        && plugin[methodName](...newArgs) === 'handled'
+    ) ? 'handled' : 'not-handled';
   };
 
   createFnHooks = (methodName, plugins) => (...args) => {
@@ -154,39 +147,37 @@ class PluginEditor extends Component {
 
     if (methodName === 'blockRendererFn') {
       let block = { props: {} };
-      for (const plugin of plugins) {
-        if (typeof plugin[methodName] !== 'function') continue;
+      plugins.forEach((plugin) => {
+        if (typeof plugin[methodName] !== 'function') return;
         const result = plugin[methodName](...newArgs);
         if (result !== undefined && result !== null) {
           const { props: pluginProps, ...pluginRest } = result; // eslint-disable-line no-use-before-define
           const { props, ...rest } = block; // eslint-disable-line no-use-before-define
           block = { ...rest, ...pluginRest, props: { ...props, ...pluginProps } };
         }
-      }
+      });
 
       return block.component ? block : false;
     } else if (methodName === 'blockStyleFn') {
       let styles;
-      for (const plugin of plugins) {
-        if (typeof plugin[methodName] !== 'function') continue;
+      plugins.forEach((plugin) => {
+        if (typeof plugin[methodName] !== 'function') return;
         const result = plugin[methodName](...newArgs);
         if (result !== undefined && result !== null) {
           styles = (styles ? (`${styles} `) : '') + result;
         }
-      }
+      });
 
       return styles || false;
     }
 
-    for (const plugin of plugins) {
-      if (typeof plugin[methodName] !== 'function') continue;
-      const result = plugin[methodName](...newArgs);
-      if (result !== undefined) {
-        return result;
-      }
-    }
-
-    return false;
+    let result;
+    const wasHandled = plugins.some((plugin) => {
+      if (typeof plugin[methodName] !== 'function') return false;
+      result = plugin[methodName](...newArgs);
+      return result !== undefined;
+    });
+    return wasHandled ? result : false;
   };
 
   createPluginHooks = () => {
@@ -263,15 +254,15 @@ class PluginEditor extends Component {
 
   resolveCustomStyleMap = () => (
     this.props.plugins
-     .filter((plug) => plug.customStyleMap !== undefined)
-     .map((plug) => plug.customStyleMap)
-     .concat([this.props.customStyleMap])
-     .reduce((styles, style) => (
-       {
-         ...styles,
-         ...style,
-       }
-     ), {})
+      .filter((plug) => plug.customStyleMap !== undefined)
+      .map((plug) => plug.customStyleMap)
+      .concat([this.props.customStyleMap])
+      .reduce((styles, style) => (
+        {
+          ...styles,
+          ...style,
+        }
+      ), {})
   );
 
   resolveblockRenderMap = () => {
@@ -290,8 +281,8 @@ class PluginEditor extends Component {
   resolveAccessibilityProps = () => {
     let accessibilityProps = {};
     const plugins = [this.props, ...this.resolvePlugins()];
-    for (const plugin of plugins) {
-      if (typeof plugin.getAccessibilityProps !== 'function') continue;
+    plugins.forEach((plugin) => {
+      if (typeof plugin.getAccessibilityProps !== 'function') return;
       const props = plugin.getAccessibilityProps();
       const popupProps = {};
 
@@ -312,7 +303,7 @@ class PluginEditor extends Component {
         ...props,
         ...popupProps,
       };
-    }
+    });
 
     return accessibilityProps;
   };