+++ draft-js-sticker-plugin/src/modifiers/addSticker.js
@@ -7,7 +7,6 @@ import {
   CharacterMetadata,
   ContentBlock,
   EditorState,
-  Entity,
   genKey,
   Modifier,
 } from 'draft-js';
@@ -54,14 +53,17 @@ export default (editorState: Object, stickerId: any) => {
   const newContentStateAfterSplit = Modifier.setBlockType(insertionTargetBlock, insertionTargetSelection, 'sticker');
 
   // creating a new ContentBlock including the entity with data
-  const entityKey = Entity.create('sticker', 'IMMUTABLE', { id: stickerId });
+  const contentStateWithEntity = newContentStateAfterSplit.createEntity(
+    'sticker', 'IMMUTABLE', { id: stickerId }
+  );
+  const entityKey = contentStateWithEntity.getLastCreatedEntityKey();
   const charDataOfSticker = CharacterMetadata.create({ entity: entityKey });
 
   const fragmentArray = [
     new ContentBlock({
       key: genKey(),
       type: 'sticker',
-      text: '',
+      text: ' ',
       characterList: List(Repeat(charDataOfSticker, 1)), // eslint-disable-line new-cap
     }),